requirements.txt
Flask
Flask-CORS
mysql-connector-python
elasticsearch
gunicorn
python-dotenv
bcrypt
requests
secrets
